+++ tests/unit/test_compute_config.py
@@ -1,185 +1,57 @@
-import unittest
-
-import dbt.exceptions
+import pytest
 from dbt.adapters.databricks import connections
 from dbt.adapters.databricks.credentials import DatabricksCredentials
-from dbt.contracts.graph import model_config
-from dbt.contracts.graph import nodes
+from dbt_common.exceptions import DbtRuntimeError
+from mock import Mock
 
 
-class TestDatabricksConnectionHTTPPath(unittest.TestCase):
+class TestDatabricksConnectionHTTPPath:
     """Test the various cases for determining a specified warehouse."""
 
-    errMsg = (
-        "Compute resource foo does not exist or does not specify http_path, " "relation: a_relation"
-    )
-
-    def test_get_http_path_model(self):
-        default_path = "my_http_path"
-        creds = DatabricksCredentials(http_path=default_path)
-
-        path = connections._get_http_path(None, creds)
-        self.assertEqual(default_path, path)
-
-        node = nodes.ModelNode(
-            relation_name="a_relation",
-            database="database",
-            schema="schema",
-            name="node_name",
-            resource_type="model",
-            package_name="package",
-            path="path",
-            original_file_path="orig_path",
-            unique_id="uniqueID",
-            fqn=[],
-            alias="alias",
-            checksum=None,
-        )
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config = model_config.ModelConfig()
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config._extra = {}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config._extra["databricks_compute"] = "foo"
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {"foo": {}}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {"foo": {"http_path": "alternate_path"}}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual("alternate_path", path)
-
-    def test_get_http_path_seed(self):
-        default_path = "my_http_path"
-        creds = DatabricksCredentials(http_path=default_path)
-
-        path = connections._get_http_path(None, creds)
-        self.assertEqual(default_path, path)
-
-        node = nodes.SeedNode(
-            relation_name="a_relation",
-            database="database",
-            schema="schema",
-            name="node_name",
-            resource_type="model",
-            package_name="package",
-            path="path",
-            original_file_path="orig_path",
-            unique_id="uniqueID",
-            fqn=[],
-            alias="alias",
-            checksum=None,
-        )
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config = model_config.SeedConfig()
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config._extra = {}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
-
-        node.config._extra["databricks_compute"] = "foo"
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {"foo": {}}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
-
-        creds.compute = {"foo": {"http_path": "alternate_path"}}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual("alternate_path", path)
-
-    def test_get_http_path_snapshot(self):
-        default_path = "my_http_path"
-        creds = DatabricksCredentials(http_path=default_path)
-
-        path = connections._get_http_path(None, creds)
-        self.assertEqual(default_path, path)
-
-        node = nodes.SnapshotNode(
-            config=None,
-            relation_name="a_relation",
-            database="database",
-            schema="schema",
-            name="node_name",
-            resource_type="model",
-            package_name="package",
-            path="path",
-            original_file_path="orig_path",
-            unique_id="uniqueID",
-            fqn=[],
-            alias="alias",
-            checksum=None,
+    @pytest.fixture(scope="class")
+    def err_msg(self):
+        return (
+            "Compute resource foo does not exist or does not specify http_path,"
+            " relation: a_relation"
         )
 
-        node.config = model_config.SnapshotConfig()
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
+    @pytest.fixture(scope="class")
+    def path(self):
+        return "my_http_path"
 
-        node.config._extra = {}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual(default_path, path)
+    @pytest.fixture
+    def creds(self, path):
+        return DatabricksCredentials(http_path=path)
 
-        node.config._extra["databricks_compute"] = "foo"
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
+    @pytest.fixture
+    def node(self):
+        n = Mock()
+        n.config = {}
+        n.relation_name = "a_relation"
+        return n
+
+    def test_get_http_path__empty(self, path, creds):
+        assert connections._get_http_path(None, creds) == path
+
+    def test_get_http_path__no_compute(self, node, path, creds):
+        assert connections._get_http_path(node, creds) == path
+
+    def test_get_http_path__missing_compute(self, node, creds, err_msg):
+        node.config["databricks_compute"] = "foo"
+        with pytest.raises(DbtRuntimeError) as exc:
             connections._get_http_path(node, creds)
 
-        creds.compute = {}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
-            connections._get_http_path(node, creds)
+        assert err_msg in str(exc.value)
 
+    def test_get_http_path__empty_compute(self, node, creds, err_msg):
+        node.config["databricks_compute"] = "foo"
         creds.compute = {"foo": {}}
-        with self.assertRaisesRegex(
-            dbt.exceptions.DbtRuntimeError,
-            self.errMsg,
-        ):
+        with pytest.raises(DbtRuntimeError) as exc:
             connections._get_http_path(node, creds)
 
+        assert err_msg in str(exc.value)
+
+    def test_get_http_path__matching_compute(self, node, creds):
+        node.config["databricks_compute"] = "foo"
         creds.compute = {"foo": {"http_path": "alternate_path"}}
-        path = connections._get_http_path(node, creds)
-        self.assertEqual("alternate_path", path)
+        assert "alternate_path" == connections._get_http_path(node, creds)