+++ tests/unit/test_utils.py
@@ -1,15 +1,14 @@
-import unittest
-
 from dbt.adapters.databricks.utils import redact_credentials
 from dbt.adapters.databricks.utils import remove_ansi
 
 
-class TestDatabricksUtils(unittest.TestCase):
-    def test_redact_credentials_copy_into(self):
+class TestDatabricksUtils:
+    def test_redact_credentials__no_credentials(self):
         sql = "copy into target_table\nfrom source_table\nfileformat = parquet"
         expected = sql
-        self.assertEqual(redact_credentials(sql), expected)
+        assert redact_credentials(sql) == expected
 
+    def test_redact_credentials__single_credential(self):
         sql = (
             "copy into target_table\n"
             "from source_table\n"
@@ -26,8 +25,10 @@ class TestDatabricksUtils(unittest.TestCase):
             "  )\n"
             "fileformat = parquet"
         )
-        self.assertEqual(redact_credentials(sql), expected)
 
+        assert redact_credentials(sql) == expected
+
+    def test_redact_credentials__multiple_credentials(self):
         sql = (
             "copy into target_table\n"
             "from source_table\n"
@@ -44,31 +45,7 @@ class TestDatabricksUtils(unittest.TestCase):
             "  )\n"
             "fileformat = parquet"
         )
-        self.assertEqual(redact_credentials(sql), expected)
-
-        sql = (
-            "copy into target_table\n"
-            "from source_table\n"
-            "  WITH (\n"
-            "    credential (\n"
-            "      'KEY1' = 'VALUE1', 'KEY2' = 'VALUE2', 'KEY3' = 'VALUE3'\n"
-            "    )\n"
-            "  )\n"
-            "fileformat = parquet\n"
-            "format_options ('mergeSchema' = 'True')\n"
-            "copy_options ('mergeSchema' = 'True')"
-        )
-        expected = (
-            "copy into target_table\n"
-            "from source_table\n"
-            "  WITH (\n"
-            "    credential ('KEY1' = '[REDACTED]', 'KEY2' = '[REDACTED]', 'KEY3' = '[REDACTED]')\n"
-            "  )\n"
-            "fileformat = parquet\n"
-            "format_options ('mergeSchema' = 'True')\n"
-            "copy_options ('mergeSchema' = 'True')"
-        )
-        self.assertEqual(redact_credentials(sql), expected)
+        assert redact_credentials(sql) == expected
 
     def test_remove_ansi(self):
         test_string = """Python model failed with traceback as:
@@ -87,4 +64,4 @@ class TestDatabricksUtils(unittest.TestCase):
        71
        72 # how to execute python model in notebook
 """
-        self.assertEqual(remove_ansi(test_string), expected_string)
+        assert remove_ansi(test_string) == expected_string